+++ R/ftirs.R
@@ -47,7 +47,7 @@ read_ftirs_file <- function(single_filepath, interpolate = TRUE, ...) {
   x <- x %>%
     mutate(sample_id = tools::file_path_sans_ext(fs::path_file(single_filepath)))
 
-   x <- as.ftirs(x)
+   x <- as_ftirs(x)
   return(x)
 }
 
@@ -73,7 +73,7 @@ read_ftirs <- function(dir_path, wet_chem_path = NULL, format = "long", ...) {
     x <- read_wet_chem(wet_chem_path, x)
   }
 
-  #x <- as.ftirs(x)
+  #x <- as_ftirs(x)
   if (format == "wide") {
     x <- pivot_wider(x)
   }
@@ -120,7 +120,7 @@ pivot_wider.ftirs <- function(ftirs_data_long, ...) {
     ) %>%
     column_to_rownames(var = "sample_id")
 
-  ftirs_data_wide <- as.ftirs(ftirs_data_wide)
+  ftirs_data_wide <- as_ftirs(ftirs_data_wide)
   return(ftirs_data_wide)
 }
 
@@ -156,7 +156,7 @@ pivot_longer.ftirs <- function(ftirs_data_wide, wet_chem, ...) {
       )
   }
 
-  ftirs_data_long <- as.ftirs(ftirs_data_long) %>%
+  ftirs_data_long <- as_ftirs(ftirs_data_long) %>%
     mutate(wavenumber = as.numeric(wavenumber))
   return(ftirs_data_long)
 }
@@ -165,7 +165,7 @@ pivot_longer.ftirs <- function(ftirs_data_wide, wet_chem, ...) {
 #' @param obj any R object
 #' @param ... Other arguments passed on to methods. Not currently used.
 #' @export
-is.ftirs <- function(obj, ...) {
+is_ftirs <- function(obj, ...) {
   "ftirs" %in% class(obj)
 }
 
@@ -173,7 +173,7 @@ is.ftirs <- function(obj, ...) {
 #' This only changes the class label of the object in order to access the methods of the class. It does not change anything about the object besides the classification.
 #' @param df A data.frame to coerce to class `ftirs`.
 #' @export
-as.ftirs <- function(df) {
+as_ftirs <- function(df) {
   if ("data.frame" %in% class(df)) {
     class(df) <- c("ftirs", class(df))
   } else {