+++ NEWS.md
@@ -1,7 +1,7 @@
 # ftirsr 0.0.1.0
 * Added `ftirs` class
 * Added methods for `ftirs` class:  `pivot_longer`, `pivot_wider`, `predict`, 
-* Added functions: `read_ftirs_file`, `read_ftirs`, `interpolate_ftirs`, `as.ftirs`, `is.ftirs`, `read_wet_chem`
+* Added functions: `read_ftirs_file`, `read_ftirs`, `interpolate_ftirs`, `as_ftirs`, `is_ftirs`, `read_wet_chem`
 * Updated `alaskaWetChemAbsorbance` to be `alaska` and to include wavenumbers
 * Updated `greenland` to include wavenumbers